+++ src/main/java/qna/config/JpaAuditingConfig.java
@@ -1,0 +1,9 @@
+package qna.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@EnableJpaAuditing
+@Configuration
+public class JpaAuditingConfig {
+}
