+++ src/test/java/qna/domain/QuestionRepositoryTest.java
@@ -8,12 +8,12 @@ import java.util.List;
 import java.util.Optional;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase.Replace;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import org.springframework.context.annotation.Import;
+import qna.config.JpaAuditingConfig;
 
 @DataJpaTest
-@AutoConfigureTestDatabase(replace = Replace.NONE)
+@Import(JpaAuditingConfig.class)
 public class QuestionRepositoryTest {
 
     @Autowired
@@ -35,7 +35,6 @@ public class QuestionRepositoryTest {
 
         Optional<Question> foundQuestion = questionRepository.findByIdAndDeletedFalse(question.getId());
 
-        assertThat(foundQuestion).isPresent();
-        assertThat(foundQuestion.get()).isEqualTo(question);
+        assertThat(foundQuestion).hasValue(question);
     }
 }
