+++ src/test/resources/application.properties
@@ -1,0 +1,9 @@
+spring.datasource.url=jdbc:h2:~/test;MODE=MySQL
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL57Dialect
+spring.datasource.username=sa
+
+spring.jpa.hibernate.ddl-auto=create
+spring.jpa.properties.hibernate.format_sql=true
+spring.jpa.show-sql=true
+spring.test.database.replace=none
+logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
